Lazily initialize livemap

+++ src/hotspot/share/gc/z/zPage.hpp
@@ -71,7 +71,7 @@ public:
 
   ZPage* clone_limited() const;
 
-  uint32_t object_max_count() const;
+  size_t object_max_count() const;
   size_t object_alignment_shift() const;
   size_t object_alignment() const;
